Refactor: Extract Integrator class into separate files
- Move Integrator class from main.cpp to src/Integrator.h/.cpp
- Add comprehensive documentation with doxygen comments
- Support both Arduino millis() and explicit timestamp parameters
- Update unit tests to use extracted class
- Add trapezoidal rule integration explanation

Benefits:
- Improved code organization and reusability
- Better separation of concerns
- Enhanced testability with explicit time parameter
- Cleaner main.cpp with focused protocol handling
