+++ platformio.ini
@@ -36,3 +36,7 @@ lib_deps =
 build_flags = 
 	-DUNIT_TEST
 	-std=c++20
+	-I test
+build_src_filter = 
+	-<*>
+	+<../test/>
